BrookTroutTH {FSA}	R Documentation

Catch-at-age for Tobin Harbor, Isle Royale Brook Trout.

Description

Catch-at-age in fyke nets from 1996-1998 for “Coaster” Brook Trout (Salvelinus fontinalis) in Tobin Harbor, Isle Royale, Lake Superior.

Format

A data frame with 7 observations on the following 2 variables.

age

A numeric vector of assigned ages

catch

A numeric vector of number of Brook Trout caught

Topic(s)

Mortality

Catch Curve

Chapman-Robson

Source

Quinlan, H.R. 1999. Biological Characteristics of Coaster Brook Trout at Isle Royale National Park, Michigan, 1996-98. U.S. Fish and Wildlife Service Ashland Fishery Resources Office report. November 1999. [Was (is?) from http://www.fws.gov/midwest/ashland/brook/biochar/biolchar.html.]

See Also

Used in catchCurve and chapmanRobson examples.

Examples

str(BrookTroutTH)
head(BrookTroutTH)
plot(log(catch)~age,data=BrookTroutTH)

[Package FSA version 0.8.23 Index]