SchnuteRichards {FSA}	R Documentation

The five-parameter growth function from Schnute and Richards (1990).

Description

The five-parameter growth function from Schnute and Richards (1990). Note that this function is slightly modified (a ‘+’ was changed to a ‘-’ so that the ‘a’ parameter will be positive) from the original in Schnute and Richards (1990)

Usage

SchnuteRichards(t, Linf = NULL, k = NULL, a = NULL, b = NULL,
  c = NULL)

Arguments

t
A numeric vector of ages over which to model growth.

Linf
Mean asymptotic length.

k
The "growth coefficient" with units of (year^(-c)).

a
A dimensionless parameter

b
A dimensionless parameter.

c
A dimensionless parameter.

Value

SchnuteRichards returns a predicted size given the provided parameter values.

IFAR Chapter

None specifically, but 12-Individual Growth is related.

Author(s)

Derek H. Ogle, derek@derekogle.com

References

Schnute, J.T. and L.J. Richards. 1990. A unified approach to the analysis of fish growth, maturity, and survivorship data. Canadian Journal of Fisheries and Aquatic Sciences. 47:24-40.

See Also

See vbFuns, GompertzFuns, RichardsFuns, logisticFuns, and Schnute for similar functionality for other models.

Examples

## See the formulae
growthFunShow("SchnuteRichards",plot=TRUE)

## Simple examples
ages <- 1:15
s1 <- SchnuteRichards(ages,Linf=100,k=0.03,a=0.01,b=0.005,c=2)
plot(s1~ages,type="l",lwd=2)

[Package FSA version 0.8.24 Index]