SpotVA1 {FSA}	R Documentation

Age and length of spot.

Description

Ages (from otoliths) and lengths of Virginia Spot (Leiostomus xanthurus).

Format

A data frame of 403 observations on the following 2 variables:

tl

Measured total lengths (in inches)

age

Ages assigned from examination of otoliths

Details

Final length measurements were simulated by adding a uniform error to the value at the beginning of the length category.

Topic(s)

Growth

von Bertalanffy

Source

Extracted from Table 1 in Chapter 8 (Spot) of the VMRC Final Report on Finfish Ageing, 2002 by the Center for Quantitative Fisheries Ecology at Old Dominion University.

See Also

Used in vbFuns, vbStarts, and nlsTracePlot examples. Also see SpotVA2 in FSAdata for related data.

Examples

str(SpotVA1)
head(SpotVA1)
plot(tl~age,data=SpotVA1)

[Package FSA version 0.8.23 Index]