All known standard weight equations.

Parameters for all known standard weight equations.

Format

A data frame with observations on the following 13 variables:

species

Species name.

units

Units of measurements. Metric uses lengths in mm and weight in grams. English uses lengths in inches and weight in pounds.

type

Type of equation (linear or quadratic).

ref

Reference quartile (75, 50, or 25).

measure

The type of length measurement used -- total length (TL) or fork length (FL).

method

The type of method used to derive the equation (RLP,EmP, or Other).

min.len

Minimum total length (mm or in, depending on units) for which the equation should be applied.

max.len

Maximum total length (mm or in, depending on units) for which the equation should be applied.

int

The intercept for the model.

slope

The slope for the linear models or the linear coefficient for the quadratic equation.

quad

The quadratic coefficient in the quadratic equations.

source

Source of the equation. These match the sources given in Neumann et al. 2012.

comment

Comments about use of equation.

Source

Most of these equations can be found in Neumann, R.M., C.S. Guy, and D.W. Willis. 2012. Length, Weight, and Associated Indices. Chapter 14 in Zale, A.V., D.L. Parrish, and T.M. Sutton, editors. Fisheries Techniques. American Fisheries Society, Bethesda, MD.

Some species were not in Neumann et al (2012) and are noted as such in the comments variable.

Details

The minimum TL for the English units were derived by rounding the converted minimum TL for the metric units to what seemed like common units (inches, half inches, or quarter inches).

Topic(s)

Relative weight

Standard weight

Condition

IFAR Chapter

8-Condition.

References

Ogle, D.H. 2016. Introductory Fisheries Analyses with R. Chapman & Hall/CRC, Boca Raton, FL.

See also

Examples

str(WSlit)

#> 'data.frame':	199 obs. of  13 variables:
#>  $ species: chr  "Aegean Chub" "African Sharptooth Catfish" "Alabama Bass" "Ankara Nase" ...
#>  $ units  : chr  "metric" "metric" "metric" "metric" ...
#>  $ type   : chr  "quadratic" "quadratic" "linear" "quadratic" ...
#>  $ ref    : int  75 75 75 75 75 75 50 75 75 25 ...
#>  $ measure: chr  "TL" "TL" "TL" "TL" ...
#>  $ method : chr  "EmP" "EmP" "RLP" "EmP" ...
#>  $ min.len: num  70 180 100 70 150 150 160 6 150 70 ...
#>  $ max.len: num  220 450 NA 240 NA NA NA NA NA 390 ...
#>  $ int    : num  -3.8 -3.67 -5.6 -10.02 -5.24 ...
#>  $ slope  : num  1.78 1.89 3.29 7.4 3.08 ...
#>  $ quad   : num  0.329 0.209 NA -0.971 NA ...
#>  $ source : chr  "Giannetto et al. (2012)" "Emiroglu et al. (2018)" "Dicenzo et al. (1995)" "Emiroglu et al. (2020)" ...
#>  $ comment: chr  "none" "only from Sakarya River Basin (Turkey)" "min.len not made clear (assumed same as Spotted Bass); same as Spotted Bass (Alabama subspecies)" "only from Turkey" ...

head(WSlit)

#>                      species  units      type ref measure method min.len
#> 1                Aegean Chub metric quadratic  75      TL    EmP      70
#> 2 African Sharptooth Catfish metric quadratic  75      TL    EmP     180
#> 3               Alabama Bass metric    linear  75      TL    RLP     100
#> 4                Ankara Nase metric quadratic  75      TL    EmP      70
#> 5            Arctic Grayling metric    linear  75      TL    RLP     150
#> 6            Arctic Grayling metric    linear  75      TL    EmP     150
#>   max.len     int  slope    quad                  source
#> 1     220  -3.801 1.7830  0.3290 Giannetto et al. (2012)
#> 2     450  -3.668 1.8850  0.2087  Emiroglu et al. (2018)
#> 3      NA  -5.598 3.2904      NA   Dicenzo et al. (1995)
#> 4     240 -10.017 7.4020 -0.9710  Emiroglu et al. (2020)
#> 5      NA  -5.241 3.0830      NA    Gilham et al. (2021)
#> 6      NA  -5.279 3.0960      NA    Gilham et al. (2021)
#>                                                                                            comment
#> 1                                                                                             none
#> 2                                                           only from Sakarya River Basin (Turkey)
#> 3 min.len not made clear (assumed same as Spotted Bass); same as Spotted Bass (Alabama subspecies)
#> 4                                                                                 only from Turkey
#> 5                                           authors note that either RLP or EmP method may be used
#> 6                                           authors note that either RLP or EmP method may be used