Extract diagonals from a matrix.

Extract diagonals from a matrix.

diags(x, which = 0, incl.labels = c("none", "row", "column"),
  val.name = "value", label.name = "label")

Arguments

x
A matrix with more than one row AND more than one column.

which
A single numeric that indicates which diagonal to extract. A value of zero extracts the main diagonal, whereas negative values extract diagonals from the upper triangle and positive values extract diagonals from the lower triangle. Diagonals further from the main diagonal have which values further from zero. If is.null(which), then a matrix of diagonal indices for which is shown.

incl.labels
A single string that indicates whether "row", "column", or no ("none") labels from x should be returned with the values on the diagonal. Will return numeric values if the labels are all diagonal, otherwise character labels are returned.

val.name
A single string to name the variable that contains the values from the diagonal in the returned data.frame.

label.name
A single string to name the variable that contains the labels in the returned data.frame (see incl.labels)

Value

A data.frame with one variable that contains the values from the chosen diagonal of x and, optionally, a second variable that contains the chosen labels for those values.

Examples

## Square numeric matrix
mat1 <- matrix(1:16,nrow=4)
colnames(mat1) <- LETTERS[1:ncol(mat1)]
rownames(mat1) <- 1:nrow(mat1)
mat1

#>   A B  C  D
#> 1 1 5  9 13
#> 2 2 6 10 14
#> 3 3 7 11 15
#> 4 4 8 12 16

diags(mat1,which=NULL)

#> Indices matrix corresponding to 'x'.
#>   A  B  C  D
#> 1 0 -1 -2 -3
#> 2 1  0 -1 -2
#> 3 2  1  0 -1
#> 4 3  2  1  0
#> 

diags(mat1)

#>   value
#> 1     1
#> 2     6
#> 3    11
#> 4    16

diags(mat1,which=-1)

#>   value
#> 1     5
#> 2    10
#> 3    15

diags(mat1,which=2)

#>   value
#> 1     3
#> 2     8

diags(mat1,incl.labels="row")

#>   label value
#> 1     1     1
#> 2     2     6
#> 3     3    11
#> 4     4    16

diags(mat1,which=2,incl.labels="row")

#>   label value
#> 1     3     3
#> 2     4     8

diags(mat1,which=2,incl.labels="col")

#>   label value
#> 1     A     3
#> 2     B     8

( tmp <- diags(mat1,which=2,incl.labels="row",val.name="Freq",label.name="age") )

#>   age Freq
#> 1   3    3
#> 2   4    8

str(tmp)

#> 'data.frame':	2 obs. of  2 variables:
#>  $ age : num  3 4
#>  $ Freq: int  3 8

## Rectangular numeric matrix
mat2 <- matrix(1:20,nrow=4)
colnames(mat2) <- LETTERS[1:ncol(mat2)]
rownames(mat2) <- 1:nrow(mat2)
mat2

#>   A B  C  D  E
#> 1 1 5  9 13 17
#> 2 2 6 10 14 18
#> 3 3 7 11 15 19
#> 4 4 8 12 16 20

diags(mat2,which=NULL)

#> Indices matrix corresponding to 'x'.
#>   A  B  C  D  E
#> 1 0 -1 -2 -3 -4
#> 2 1  0 -1 -2 -3
#> 3 2  1  0 -1 -2
#> 4 3  2  1  0 -1
#> 

diags(mat2,which=-1,incl.labels="row")

#>   label value
#> 1     1     5
#> 2     2    10
#> 3     3    15
#> 4     4    20

diags(mat2,which=2,incl.labels="row")

#>   label value
#> 1     3     3
#> 2     4     8

diags(mat2,which=-4,incl.labels="col")

#>   label value
#> 1     E    17

## Rectangular character matrix
mat3 <- matrix(LETTERS[1:24],nrow=3)
colnames(mat3) <- letters[1:ncol(mat3)]
rownames(mat3) <- 1:nrow(mat3)
mat3

#>   a   b   c   d   e   f   g   h  
#> 1 "A" "D" "G" "J" "M" "P" "S" "V"
#> 2 "B" "E" "H" "K" "N" "Q" "T" "W"
#> 3 "C" "F" "I" "L" "O" "R" "U" "X"

diags(mat3,which=NULL)

#> Indices matrix corresponding to 'x'.
#>   a  b  c  d  e  f  g  h
#> 1 0 -1 -2 -3 -4 -5 -6 -7
#> 2 1  0 -1 -2 -3 -4 -5 -6
#> 3 2  1  0 -1 -2 -3 -4 -5
#> 

diags(mat3,which=-1,incl.labels="row")

#>   label value
#> 1     1     D
#> 2     2     H
#> 3     3     L

diags(mat3,which=2,incl.labels="row")

#>   label value
#> 1     3     C

diags(mat3,which=-4,incl.labels="col")

#>   label value
#> 1     e     M
#> 2     f     Q
#> 3     g     U