mapvalues {FSA}	R Documentation

Replace specified values in a vector or factor with new values.

Description

Same function as in plyr. Imported by FSA and exported to minimize conflicts with other commonly used packages. See mapvalues for documentation.

[Package FSA version 0.8.24 Index]