Estimate natural mortality from a variety of empirical methods.

Several methods can be used to estimated natural mortality (M) from other types of data, including parameters from the von Bertalanffy growth equation, maximum age, and temperature. These relationships have been developed from meta-analyses of a large number of populations. Several of these methods are implemented in this function.

Mmethods(what = c("all", "tmax", "K", "Hoenig", "Pauly"))

metaM(
  method = Mmethods(),
  justM = TRUE,
  tmax = NULL,
  K = NULL,
  Linf = NULL,
  t0 = NULL,
  b = NULL,
  L = NULL,
  Temp = NULL,
  t50 = NULL,
  Winf = NULL
)

# S3 method for metaM
print(x, digits = 4, ...)

Arguments

what

A string that indicates what grouping of methods to return. Defaults to returning all methods.

method

A string that indicates which method or equation to use. See details.

justM

A logical that indicates whether just the estimate of M (TRUE; Default) or a more descriptive list should be returned.

tmax

The maximum age for the population of fish.

K

The Brody growth coefficient from the fit of the von Bertalanffy growth function.

Linf

The asymptotic mean length (cm) from the fit of the von Bertalanffy growth function.

t0

The x-intercept from the fit of the von Bertalanffy growth function.

b

The exponent from the weight-length relationship (slope from the logW-logL relationship).

L

The body length of the fish (cm).

Temp

The temperature experienced by the fish (C).

t50

The age (time) when half the fish in the population are mature.

Winf

The asymptotic mean weight (g) from the fit of the von Bertalanffy growth function.

x

A metaM object returned from metaM when justM=FALSE.

digits

A numeric that controls the number of digits printed for the estimate of M.

...

Additional arguments for methods. Not implemented.

Value

Mmethods returns a character vector with a list of methods. If only one method is chosen then metaM returns a single numeric if justM=TRUE or, otherwise, a metaM object that is a list with the following items:

method: The name for the method within the function (as given in method).

name: A more descriptive name for the method.

givens: A vector of values required by the method to estimate M.

M: The estimated natural mortality rate.

If multiple methods are chosen then a data.frame is returned with the method name abbreviation in the method variable and the associated estimated M in the M variable.

Details

One of several methods is chosen with method. The available methods can be seen with Mmethods() and are listed below with a brief description of where the equation came from. The sources (listed below) should be consulted for more specific information.

method="HoenigNLS": The “modified Hoenig equation derived with a non-linear model” as described in Then et al. (2015) on the third line of Table 3. This method was the preferred method suggested by Then et al. (2015). Requires only tmax.

method="PaulyLNoT": The “modified Pauly length equation” as described on the sixth line of Table 3 in Then et al. (2015). Then et al. (2015) suggested that this is the preferred model if maximum age (tmax) information was not available. Requires K and Linf.

method="PaulyL": The “Pauly (1980) equation using fish lengths” from his equation 11. This is the most commonly used method in the literature. Note that Pauly used common logarithms as used here but the model is often presented in other sources with natural logarithms. Requires K, Linf, and T.

method="PaulyW": The “Pauly (1980) equation for weights” from his equation 10. Requires K, Winf, and T.

method="HoeingO", method="HoeingOF", method="HoeingOM", method="HoeingOC": The original “Hoenig (1983) composite”, “fish”, “mollusc”, and “cetacean” (fit with OLS) equations from the second column on page 899 of Hoenig (1983). Requires only tmax.

method="HoeingO2", method="HoeingO2F", method="HoeingO2M", method="HoeingO2C": The original “Hoenig (1983) composite”, “fish”, “mollusc”, and “cetacean” (fit with Geometric Mean Regression) equations from the second column on page 537 of Kenchington (2014). Requires only tmax.

method="HoenigLM": The “modified Hoenig equation derived with a linear model” as described in Then et al. (2015) on the second line of Table 3. Requires only tmax.

method="HewittHoenig": The “Hewitt and Hoenig (2005) equation” from their equation 8. Requires only tmax.

method="tmax1": The “one-parameter tmax equation” from the first line of Table 3 in Then et al. (2015). Requires only tmax.

method="K1": The “one-parameter K equation” from the fourth line of Table 3 in Then et al. (2015). Requires only K.

method="K2": The “two-parameter K equation” from the fifth line of Table 3 in Then et al. (2015). Requires only K.

method="JensenK1": The “Jensen (1996) one-parameter K equation”. Requires only K.

method="JensenK2": The “Jensen (2001) two-parameter K equation” from their equation 8. Requires only K.

method="Gislason": The “Gislason et al. (2010) equation” from their equation 2. Requires K, Linf, and L.

method="AlversonCarney": The “Alverson and Carney (1975) equation” as given in equation 10 of Zhang and Megrey (2006). Requires tmax and K.

method="Charnov": The “Charnov et al. (2013) equation” as given in the second column of page 545 of Kenchington (2014). Requires K, Linf, and L.

method="ZhangMegreyD", method="ZhangMegreyP": The “Zhang and Megrey (2006) equation” as given in their equation 8 but modified for demersal or pelagic fish. Thus, the user must choose the fish type with group. Requires tmax, K, t0, t50, and b.

method="RikhterEfanov1": The “Rikhter and Efanov (1976) equation (#2)” as given in the second column of page 541 of Kenchington (2014) and in Table 6.4 of Miranda and Bettoli (2007). Requires only t50.

method="RikhterEfanov2": The “Rikhter and Efanov (1976) equation (#1)” as given in the first column of page 541 of Kenchington (2014). Requires t50, K, t0, and b.

Testing

Kenchington (2014) provided life history parameters for several stocks and used many models to estimate M. I checked the calculations for the PaulyL, PaulyW, HoenigO for Hgroup="all" and Hgroup="fish", HoenigO2 for Hgroup="all" and Hgroup="fish", "JensenK1", "Gislason", "AlversonCarney", "Charnov", "ZhangMegrey", "RikhterEfanov1", and "RikhterEfanov2" methods for three stocks. All results perfectly matched Kenchington's results for Chesapeake Bay Anchovy and Rio Formosa Seahorse. For the Norwegian Fjord Lanternfish, all results perfectly matched Kenchington's results except for when Hgroup="fish" for both HoenigO and HoenigO2.

Results for the Rio Formosa Seahorse data were also tested against results from M.empirical from fishmethods for the PaulyL, PaulyW, HoenigO for Hgroup="all" and Hgroup="fish", "Gislason", and "AlversonCarney" methods (the only methods in common between the two packages). All results matched perfectly.

IFAR Chapter

11-Mortality.

References

Ogle, D.H. 2016. Introductory Fisheries Analyses with R. Chapman & Hall/CRC, Boca Raton, FL.

Alverson, D.L. and M.J. Carney. 1975. A graphic review of the growth and decay of population cohorts. Journal du Conseil International pour l'Exploration de la Mer. 36:133-143.

Charnov, E.L., H. Gislason, and J.G. Pope. 2013. Evolutionary assembly rules for fish life histories. Fish and Fisheries. 14:213-224.

Gislason, H., N. Daan, J.C. Rice, and J.G. Pope. 2010. Size, growth, temperature and the natural mortality of marine fish. Fish and Fisheries 11:149-158.

Hewitt, D.A. and J.M. Hoenig. 2005. Comparison of two approaches for estimating natural mortality based on longevity. Fishery Bulletin. 103:433-437. [Was (is?) from http://fishbull.noaa.gov/1032/hewitt.pdf.]

Hoenig, J.M. 1983. Empirical use of longevity data to estimate mortality rates. Fishery Bulletin. 82:898-903. [Was (is?) from http://www.afsc.noaa.gov/REFM/age/Docs/Hoenig_EmpiricalUseOfLongevityData.pdf.]

Jensen, A.L. 1996. Beverton and Holt life history invariants result from optimal trade-off of reproduction and survival. Canadian Journal of Fisheries and Aquatic Sciences. 53:820-822. [Was (is?) from .]

Jensen, A.L. 2001. Comparison of theoretical derivations, simple linear regressions, multiple linear regression and principal components for analysis of fish mortality, growth and environmental temperature data. Environometrics. 12:591-598. [Was (is?) from http://deepblue.lib.umich.edu/bitstream/handle/2027.42/35236/487_ftp.pdf.]

Kenchington, T.J. 2014. Natural mortality estimators for information-limited fisheries. Fish and Fisheries. 14:533-562.

Pauly, D. 1980. On the interrelationships between natural mortality, growth parameters, and mean environmental temperature in 175 fish stocks. Journal du Conseil International pour l'Exploration de la Mer. 39:175-192. [Was (is?) from http://innri.unuftp.is/pauly/On%20the%20interrelationships%20betwe.pdf.]

Rikhter, V.A., and V.N. Efanov. 1976. On one of the approaches for estimating natural mortality in fish populations (in Russian). ICNAF Research Document 76/IV/8, 12pp.

Then, A.Y., J.M. Hoenig, N.G. Hall, and D.A. Hewitt. 2015. Evaluating the predictive performance of empirical estimators of natural mortality rate using information on over 200 fish species. ICES Journal of Marine Science. 72:82-92.

Zhang, C-I and B.A. Megrey. 2006. A revised Alverson and Carney model for estimating the instantaneous rate of natural mortality. Transactions of the American Fisheries Society. 135-620-633. [Was (is?) from http://www.pmel.noaa.gov/foci/publications/2006/zhan0531.pdf.]

See also

Author

Derek H. Ogle, derek@derekogle.com

Examples

## List names for available methods
Mmethods()
#>  [1] "HoenigNLS"      "HoenigO"        "HoenigOF"       "HoenigOM"      
#>  [5] "HoenigOC"       "HoenigO2"       "HoenigO2F"      "HoenigO2M"     
#>  [9] "HoenigO2C"      "HoenigLM"       "HewittHoenig"   "tmax1"         
#> [13] "PaulyLNoT"      "PaulyL"         "PaulyW"         "K1"            
#> [17] "K2"             "JensenK1"       "JensenK2"       "Gislason"      
#> [21] "AlversonCarney" "Charnov"        "ZhangMegreyD"   "ZhangMegreyP"  
#> [25] "RikhterEfanov1" "RikhterEfanov2"
Mmethods("tmax")
#>  [1] "tmax1"        "HoenigNLS"    "HoenigO"      "HoenigOF"     "HoenigOM"    
#>  [6] "HoenigOC"     "HoenigO2"     "HoenigO2F"    "HoenigO2M"    "HoenigO2C"   
#> [11] "HoenigLM"     "HewittHoenig"

## Simple Examples
metaM("tmax",tmax=20)
#> [1] 0.25545
metaM("tmax",tmax=20,justM=FALSE)
#> M=0.2554 as estimated with Then et al. (2015) tmax equation
#>   with givens: tmax=20
metaM("HoenigNLS",tmax=20)
#> [1] 0.3150387
metaM("HoenigNLS",tmax=20,justM=FALSE)
#> M=0.315 as estimated with Then et al. (2015) Hoenig (NLS) equation
#>   with givens: tmax=20
 
## Example Patagonian Sprat ... from Table 2 in Cerna et al. (2014)
## http://www.scielo.cl/pdf/lajar/v42n3/art15.pdf
Temp <- 11
Linf <- 17.71
K <- 0.78
t0 <- -0.46
tmax <- t0+3/K
t50 <- t0-(1/K)*log(1-13.5/Linf)
metaM("RikhterEfanov1",t50=t50)
#> [1] 1.050009
metaM("PaulyL",K=K,Linf=Linf,Temp=Temp)
#> [1] 1.14058
metaM("PaulyL",K=K,Linf=Linf,Temp=Temp,justM=FALSE)
#> M=1.1406 as estimated with Pauly (1980) length equation
#>   with givens: K=0.78, Linf=17.71, Temp=11
metaM("HoenigNLS",tmax=tmax)
#> [1] 1.602862
metaM("HoenigO",tmax=tmax)
#> [1] 1.274125
metaM("HewittHoenig",tmax=tmax)
#> [1] 1.246252
metaM("AlversonCarney",K=K,tmax=tmax)
#> [1] 1.35398

## Example of multiple calculations
metaM(c("RikhterEfanov1","PaulyL","HoenigO","HewittHoenig","AlversonCarney"),
     K=K,Linf=Linf,Temp=Temp,tmax=tmax,t50=t50)
#>           method        M
#> 1 RikhterEfanov1 1.050009
#> 2         PaulyL 1.140580
#> 3        HoenigO 1.274125
#> 4   HewittHoenig 1.246252
#> 5 AlversonCarney 1.353980

## Example of multiple methods using Mmethods
# select some methods
metaM(Mmethods()[-c(15,20,22:24,26)],K=K,Linf=Linf,Temp=Temp,tmax=tmax,t50=t50)
#>            method         M
#> 1       HoenigNLS 1.6028619
#> 2         HoenigO 1.2741252
#> 3        HoenigOF 1.2562214
#> 4        HoenigOM 1.2401272
#> 5        HoenigOC 0.8835623
#> 6        HoenigO2 1.4786176
#> 7       HoenigO2F 1.5784652
#> 8       HoenigO2M 1.4266652
#> 9       HoenigO2C 1.4625421
#> 10       HoenigLM 1.6243510
#> 11   HewittHoenig 1.2462517
#> 12          tmax1 1.5087915
#> 13      PaulyLNoT 1.3304645
#> 14         PaulyL 1.1405804
#> 15             K1 1.3197600
#> 16             K2 1.3070000
#> 17       JensenK1 1.1700000
#> 18       JensenK2 1.3566000
#> 19 AlversonCarney 1.3539801
#> 20 RikhterEfanov1 1.0500095
# select just the Hoenig methods
metaM(Mmethods("Hoenig"),K=K,Linf=Linf,Temp=Temp,tmax=tmax,t50=t50)
#>          method         M
#> 1     HoenigNLS 1.6028619
#> 2       HoenigO 1.2741252
#> 3      HoenigOF 1.2562214
#> 4      HoenigOM 1.2401272
#> 5      HoenigOC 0.8835623
#> 6      HoenigO2 1.4786176
#> 7     HoenigO2F 1.5784652
#> 8     HoenigO2M 1.4266652
#> 9     HoenigO2C 1.4625421
#> 10     HoenigLM 1.6243510
#> 11 HewittHoenig 1.2462517