Fisheries stock assessment methods and data.

Functions to support basic fisheries stock assessment methods.

Details

Functions from this package can be used to perform a variety of basic fisheries stock assessment methods. Detailed descriptions for most functions are available in the Introductory Fisheries Analysis with R book (Ogle 2016). Vignettes for the boxed examples in the “Analysis and Interpretation of Freshwater Fisheries Data” book can be viewed with fishR("AIFFD").

Questions, comments, or suggestions should be given on the GitHub FSA Issues page.

Packages with related functionality by the same author are

The FSAdata package contains additional data sets.

The FSAsim package simulation routines for various fisheries methods.

The FSAWs package contains functions for developing and validating standard weight equations.

References

Ogle, D.H. 2016. Introductory Fisheries Analyses with R. Chapman & Hall/CRC, Boca Raton, FL.