Mirex {FSA}	R Documentation

Mirex concentration, weight, capture year, and species of Lake Ontario salmon.

Description

Mirex concentration, weight, capture year, and species of Lake Ontario Coho and Chinook salmon.

Format

A data frame with 122 observations on the following 4 variables.

year

a numeric vector of capture years

weight

a numeric vector of salmon weights (kg)

mirex

a numeric vector of mirex concentration in the salmon tissue (mg/kg)

species

a factor with levels chinook and coho

Details

The year variable should be converted to a factor as shown in the example.

Topic(s)

Linear models

Other

Source

From (actual data) Makarewicz, J.C., E.Damaske, T.W. Lewis, and M. Merner. 2003. Trend analysis reveals a recent reduction in mirex concentrations in Coho (Oncorhynchus kisutch) and Chinook (O. tshawytscha) Salmon from Lake Ontario. Environmental Science and Technology, 37:1521-1527.

See Also

Used in fitPlot, residPlot, compSlopes, compIntercepts, hoCoef, and rSquared examples.

Examples

Mirex$year <- factor(Mirex$year)
lm1 <- lm(mirex~weight*year*species,data=Mirex)
anova(lm1)

[Package FSA version 0.8.25 Index]