PikeNY {FSA}	R Documentation

Summarized multiple mark-recapture data for all Northern Pike from Buckhorn Marsh, NY.

Description

Summary results of capture histories (number captured, number of recaptured fish, and number of unmarked fish that were marked) for all Buckhorn Marsh Northern Pike (Esox lucius).

Format

A data frame with 21 observations on the following 4 variables:

date

Capture date

n

Total fish captured in each sample

m

Marked fish captured in each sample

R

Marked fish returned to the population

Topic(s)

Population Size

Abundance

Mark-Recapture

Capture-Recapture

Schnabel

Schumacher-Eschmeyer

Source

New York Power Authority. 2004. Use of Buckhorn Marsh and Grand Island tributaries by Northern Pike for spawning and as a nursery. Technical report, New York Power Authority, January 2004. Niagara Power Project (FERC No. 2216).

See Also

Used in mrClosed examples. Also see PikeNYPartial1.

Examples

str(PikeNY)
head(PikeNY)

[Package FSA version 0.8.25 Index]