SMBassWB {FSA}	R Documentation

Growth increment data for West Bearskin Lake, MN, Smallmouth Bass.

Description

Growth data from Smallmouth Bass (Micropterus dolomieu) captured in West Bearskin Lake, MN. Five samples were collected over three years (1988-1990) with two gears (fall – trapnets, spring – electrofishing).

Format

A data frame of 445 observations on the following 20 variables:

species

Species of the fish (SMB for each fish in this file)

lake

Lake fish was captured in (WB for each fish in this file)

gear

Gear used to capture the fish (T=Trapnet and E=Electrofishing)

yearcap

Year fish was captured (1988, 1989, or 1990)

fish

A unique identifier for each fish

agecap

Assigned age-at-capture for the fish (from scales)

lencap

Total length-at-capture for the fish (mm)

anu1

Magnified scale radius (mm) to the 1st annulus

anu2

Magnified scale radius (mm) to the 2nd annulus

anu3

Magnified scale radius (mm) to the 3rd annulus

anu4

Magnified scale radius (mm) to the 4th annulus

anu5

Magnified scale radius (mm) to the 5th annulus

anu6

Magnified scale radius (mm) to the 6th annulus

anu7

Magnified scale radius (mm) to the 7th annulus

anu8

Magnified scale radius (mm) to the 8th annulus

anu9

Magnified scale radius (mm) to the 9th annulus

anu10

Magnified scale radius (mm) to the 10th annulus

anu11

Magnified scale radius (mm) to the 11th annulus

anu12

Magnified scale radius (mm) to the 12th annulus

radcap

Total scale radius at time of capture

Topic(s)

Growth increment analysis

Weisberg linear growth model

Back-Calculation

Note

Data are in one-fish-per-line format.

Source

Data from the linear growth modeling software distributed in support of Weisberg, S. 1993. Using hard-part increment data to estimate age and environmental effects. Canadian Journal of Fisheries and Aquatic Sciences 50:1229-1237.

See Also

Used in capHistSum and mrClosed examples. Also see wblake from alr4 for the same dataset with only the agecap, lencap, and radcap variables.

Examples

str(SMBassWB)
head(SMBassWB)

[Package FSA version 0.8.25 Index]