Summary statistics for a numeric variable.

Summary statistics for a single numeric variable, possibly separated by the levels of a factor variable or variables. This function is very similar to summary for a numeric variable.

Summarize(object, ...)

# S3 method for default
Summarize(
  object,
  digits = getOption("digits"),
  na.rm = TRUE,
  exclude = NULL,
  nvalid = c("different", "always", "never"),
  percZero = c("different", "always", "never"),
  ...
)

# S3 method for formula
Summarize(
  object,
  data = NULL,
  digits = getOption("digits"),
  na.rm = TRUE,
  exclude = NULL,
  nvalid = c("different", "always", "never"),
  percZero = c("different", "always", "never"),
  ...
)

Arguments

object

A vector of numeric data.

...

Not implemented.

digits

A single numeric that indicates the number of decimals to round the numeric summaries.

na.rm

A logical that indicates whether numeric missing values (NA) should be removed (=TRUE, default) or not.

exclude

A string that contains the level that should be excluded from a factor variable.

nvalid

A string that indicates how the “validn” result will be handled. If "always" then “validn” will always be shown and if "never" then “validn” will never be shown. However, if "different" (DEFAULT), then “validn” will only be shown if it differs from “n” (or if at least one group differs from “n” when summarized by multiple groups).

percZero

A string that indicates how the “percZero” result will be handled. If "always" then “percZero” will always be shown and if "never" then “percZero” will never be shown. However, if "different" (DEFAULT), then “percZero” will only be shown if it is greater than zero (or if at least one group is greater than zero when summarized by multiple groups).

data

A data.frame that contains the variables in formula.

Value

A named vector or data frame (when a quantitative variable is separated by one or two factor variables) of summary statistics for numeric data.

Details

This function is primarily used with formulas of the following types (where quant and factor generically represent quantitative/numeric and factor variables, respectively):

Formula	Description of Summary

~quant	Numerical summaries (see below) of quant.

quant~factor	Summaries of quant separated by levels in factor.

quant~factor1*factor2	Summaries of quant separated by the combined levels in factor1 and factor2.

Numerical summaries include all results from summary (min, Q1, mean, median, Q3, and max) and the sample size, valid sample size (sample size minus number of NAs), and standard deviation (i.e., sd). NA values are removed from the calculations with na.rm=TRUE (the DEFAULT). The number of digits in the returned results are controlled with digits=.

Note

Students often need to examine basic statistics of a quantitative variable separated for different levels of a categorical variable. These results may be obtained with tapply, by, or aggregate (or with functions in other packages), but the use of these functions is not obvious to newbie students or return results in a format that is not obvious to newbie students. Thus, the formula method to Summarize allows newbie students to use a common notation (i.e., formula) to easily compute summary statistics for a quantitative variable separated by the levels of a factor.

See also

Author

Derek H. Ogle, derek@derekogle.com

Examples

## Create a data.frame of "data"
n <- 102
d <- data.frame(y=c(0,0,NA,NA,NA,runif(n-5)),
                w=sample(7:9,n,replace=TRUE),
                v=sample(0:2,n,replace=TRUE),
                g1=factor(sample(c("A","B","C",NA),n,replace=TRUE)),
                g2=factor(sample(c("male","female","UNKNOWN"),n,replace=TRUE)),
                g3=sample(c("a","b","c","d"),n,replace=TRUE),
                stringsAsFactors=FALSE)

# typical output of summary() for a numeric variable
summary(d$y)   
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max.    NA's 
#>  0.0000  0.2643  0.5103  0.4818  0.7114  0.9730       3 

# this function           
Summarize(d$y,digits=3)
#>        n   nvalid     mean       sd      min       Q1   median       Q3 
#>  102.000   99.000    0.482    0.276    0.000    0.264    0.510    0.711 
#>      max percZero 
#>    0.973    2.020 
Summarize(~y,data=d,digits=3)
#>        n   nvalid     mean       sd      min       Q1   median       Q3 
#>  102.000   99.000    0.482    0.276    0.000    0.264    0.510    0.711 
#>      max percZero 
#>    0.973    2.020 
Summarize(y~1,data=d,digits=3)
#>        n   nvalid     mean       sd      min       Q1   median       Q3 
#>  102.000   99.000    0.482    0.276    0.000    0.264    0.510    0.711 
#>      max percZero 
#>    0.973    2.020 

# note that nvalid is not shown if there are no NAs and
#   percZero is not shown if there are no zeros
Summarize(~w,data=d,digits=3)
#>       n    mean      sd     min      Q1  median      Q3     max 
#> 102.000   7.931   0.799   7.000   7.000   8.000   9.000   9.000 
Summarize(~v,data=d,digits=3)
#>        n     mean       sd      min       Q1   median       Q3      max 
#>  102.000    1.127    0.875    0.000    0.000    1.000    2.000    2.000 
#> percZero 
#>   32.353 

# note that the nvalid and percZero results can be forced to be shown
Summarize(~w,data=d,digits=3,nvalid="always",percZero="always")
#>        n   nvalid     mean       sd      min       Q1   median       Q3 
#>  102.000  102.000    7.931    0.799    7.000    7.000    8.000    9.000 
#>      max percZero 
#>    9.000    0.000 

## Numeric vector by levels of a factor variable
Summarize(y~g1,data=d,digits=3)
#>   g1  n nvalid  mean    sd   min    Q1 median    Q3   max percZero
#> 1  A 17     16 0.419 0.304 0.012 0.230  0.344 0.649 0.973    0.000
#> 2  B 24     23 0.401 0.258 0.023 0.181  0.372 0.581 0.903    0.000
#> 3  C 37     37 0.549 0.257 0.000 0.376  0.571 0.736 0.962    2.703
Summarize(y~g2,data=d,digits=3)
#>        g2  n nvalid  mean    sd   min    Q1 median    Q3   max percZero
#> 1 UNKNOWN 25     23 0.555 0.258 0.001 0.420  0.555 0.715 0.973        0
#> 2  female 41     40 0.424 0.269 0.000 0.246  0.377 0.663 0.903        5
#> 3    male 36     36 0.499 0.287 0.023 0.244  0.533 0.762 0.895        0
Summarize(y~g2,data=d,digits=3,exclude="UNKNOWN")
#>       g2  n nvalid  mean    sd   min    Q1 median    Q3   max percZero
#> 1 female 41     40 0.424 0.269 0.000 0.246  0.377 0.663 0.903        5
#> 2   male 36     36 0.499 0.287 0.023 0.244  0.533 0.762 0.895        0

## Numeric vector by levels of two factor variables
Summarize(y~g1+g2,data=d,digits=3)
#>   g1      g2  n nvalid  mean    sd   min    Q1 median    Q3   max percZero
#> 1  A UNKNOWN  5      4 0.604 0.358 0.182 0.378  0.629 0.855 0.973    0.000
#> 2  B UNKNOWN  4      4 0.532 0.095 0.426 0.479  0.526 0.579 0.650    0.000
#> 3  C UNKNOWN 11     11 0.573 0.309 0.001 0.406  0.672 0.746 0.962    0.000
#> 4  A  female  7      7 0.254 0.251 0.012 0.047  0.262 0.344 0.721    0.000
#> 5  B  female 13     12 0.404 0.250 0.080 0.251  0.360 0.457 0.903    0.000
#> 6  C  female 12     12 0.503 0.273 0.000 0.270  0.598 0.722 0.844    8.333
#> 7  A    male  5      5 0.504 0.257 0.246 0.273  0.510 0.625 0.863    0.000
#> 8  B    male  7      7 0.319 0.326 0.023 0.098  0.156 0.491 0.877    0.000
#> 9  C    male 14     14 0.568 0.206 0.193 0.475  0.549 0.744 0.861    0.000
Summarize(y~g1+g2,data=d,digits=3,exclude="UNKNOWN")
#>   g1     g2  n nvalid  mean    sd   min    Q1 median    Q3   max percZero
#> 1  A female  7      7 0.254 0.251 0.012 0.047  0.262 0.344 0.721    0.000
#> 2  B female 13     12 0.404 0.250 0.080 0.251  0.360 0.457 0.903    0.000
#> 3  C female 12     12 0.503 0.273 0.000 0.270  0.598 0.722 0.844    8.333
#> 4  A   male  5      5 0.504 0.257 0.246 0.273  0.510 0.625 0.863    0.000
#> 5  B   male  7      7 0.319 0.326 0.023 0.098  0.156 0.491 0.877    0.000
#> 6  C   male 14     14 0.568 0.206 0.193 0.475  0.549 0.744 0.861    0.000

## What happens if RHS of formula is not a factor
Summarize(y~w,data=d,digits=3)
#>   w  n nvalid  mean    sd   min    Q1 median    Q3   max percZero
#> 1 7 36     35 0.435 0.270 0.001 0.181  0.426 0.663 0.973    0.000
#> 2 8 37     37 0.524 0.290 0.000 0.259  0.542 0.760 0.957    2.703
#> 3 9 29     27 0.484 0.261 0.000 0.322  0.497 0.662 0.962    3.704

## Summarizing multiple variables in a data.frame (must reduce to numerics)
lapply(as.list(d[,1:3]),Summarize,digits=4)
#> $y
#>        n   nvalid     mean       sd      min       Q1   median       Q3 
#> 102.0000  99.0000   0.4818   0.2755   0.0000   0.2643   0.5103   0.7114 
#>      max percZero 
#>   0.9730   2.0202 
#> 
#> $w
#>        n     mean       sd      min       Q1   median       Q3      max 
#> 102.0000   7.9314   0.7993   7.0000   7.0000   8.0000   9.0000   9.0000 
#> 
#> $v
#>        n     mean       sd      min       Q1   median       Q3      max 
#> 102.0000   1.1275   0.8751   0.0000   0.0000   1.0000   2.0000   2.0000 
#> percZero 
#>  32.3529 
#>