Ages and lengths for a hypothetical sample from Westerheim and Ricker (1979).

Ages and lengths for a hypothetical sample in Westerheim and Ricker (1979).

Format

A data frame of 2369 observations on the following 3 variables:

ID

Unique fish identifiers

len

Length of an individual fish

age

Age of an individual fish

Source

Simulated from Table 2A in Westerheim, S.J. and W.E. Ricker. 1979. Bias in using age-length key to estimate age-frequency distributions. Journal of the Fisheries Research Board of Canada. 35:184-189. CSV file

Details

Age-length data in 5-cm increments taken exactly from Table 2A of the source which was a sample from a hypothetical population in which year-class strength varied in the ratio 2:1 and the rate of increase in length decreased with age. Actual lengths in each 5-cm interval were simulated with a uniform distribution. The aged fish in this file were randomly selected and an assessed age was assigned according to the information in Table 2A.

Topic(s)

Age-Length Key

Examples

str(WR79)
#> 'data.frame':	2369 obs. of  3 variables:
#>  $ ID : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ len: int  37 37 39 37 37 35 42 42 42 44 ...
#>  $ age: int  NA NA NA NA 4 4 NA NA NA NA ...
head(WR79)
#>   ID len age
#> 1  1  37  NA
#> 2  2  37  NA
#> 3  3  39  NA
#> 4  4  37  NA
#> 5  5  37   4
#> 6  6  35   4

## Extract the aged sample
WR79.aged <- subset(WR79,!is.na(age))
str(WR79.aged)
#> 'data.frame':	203 obs. of  3 variables:
#>  $ ID : int  5 6 21 32 40 57 59 70 94 117 ...
#>  $ len: int  37 35 42 43 40 41 44 46 45 47 ...
#>  $ age: int  4 4 4 4 4 4 4 4 4 4 ...

## Extract the length sample
WR79.length <- subset(WR79,is.na(age))
str(WR79.length)
#> 'data.frame':	2166 obs. of  3 variables:
#>  $ ID : int  1 2 3 4 7 8 9 10 11 12 ...
#>  $ len: int  37 37 39 37 42 42 42 44 44 43 ...
#>  $ age: int  NA NA NA NA NA NA NA NA NA NA ...