All known standard weight equations.

Parameters for all known standard weight equations.

Format

A data frame with observations on the following 13 variables:

species

Species name.

units

Units of measurements. Metric uses lengths in mm and weight in grams. English uses lengths in inches and weight in pounds.

type

Type of equation (linear or quadratic).

ref

Reference quartile (75, 50, or 25).

measure

The type of length measurement used -- total length (TL) or fork length (FL).

method

The type of method used to derive the equation (RLP,EmP, or Other).

min.len

Minimum total length (mm or in, depending on units) for which the equation should be applied.

max.len

Maximum total length (mm or in, depending on units) for which the equation should be applied.

int

The intercept for the model.

slope

The slope for the linear models or the linear coefficient for the quadratic equation.

quad

The quadratic coefficient in the quadratic equations.

source

Source of the equation. These match the sources given in Neumann et al. 2012.

comment

Comments about use of equation.

Source

Most of these equations can be found in Neumann, R.M., C.S. Guy, and D.W. Willis. 2012. Length, Weight, and Associated Indices. Chapter 14 in Zale, A.V., D.L. Parrish, and T.M. Sutton, editors. Fisheries Techniques. American Fisheries Society, Bethesda, MD.

Some species were not in Neumann et al (2012) and are noted as such in the comments variable.

Details

The minimum TL for the English units were derived by rounding the converted minimum TL for the metric units to what seemed like common units (inches, half inches, or quarter inches).

Topic(s)

Relative weight

Standard weight

Condition

IFAR Chapter

8-Condition.

References

Ogle, D.H. 2016. Introductory Fisheries Analyses with R. Chapman & Hall/CRC, Boca Raton, FL.

See also

Examples

str(WSlit)

#> 'data.frame':	179 obs. of  13 variables:
#>  $ species: Factor w/ 93 levels "Aegean Chub",..: 1 2 2 3 3 3 4 4 4 5 ...
#>  $ units  : Factor w/ 2 levels "English","metric": 2 1 2 2 2 2 2 2 2 1 ...
#>  $ type   : Factor w/ 2 levels "linear","quadratic": 2 1 1 2 2 2 2 2 2 1 ...
#>  $ ref    : int  75 75 75 25 50 75 25 50 75 75 ...
#>  $ measure: Factor w/ 4 levels "BL","CL","FL",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ method : Factor w/ 3 levels "EmP","Other",..: 1 3 3 1 1 1 1 1 1 3 ...
#>  $ min.len: num  70 6 150 70 70 70 40 40 40 5 ...
#>  $ max.len: num  220 NA NA 390 390 390 390 390 390 NA ...
#>  $ int    : num  -3.8 -3.35 -5.07 -6 -5.4 ...
#>  $ slope  : num  1.78 3.12 3.12 3.63 3.15 ...
#>  $ quad   : num  0.329 NA NA -0.111 -0.011 0.188 -0.128 -0.043 0.075 NA ...
#>  $ source : Factor w/ 49 levels "Anderson and Gutreuter (1983)",..: 18 5 5 11 11 11 11 11 11 5 ...
#>  $ comment: Factor w/ 13 levels "body length",..: 7 6 6 7 11 7 7 11 11 6 ...

head(WSlit)

#>                   species   units      type ref measure method min.len max.len
#> 1             Aegean Chub  metric quadratic  75      TL    EmP      70     220
#> 2        Bigmouth Buffalo English    linear  75      TL    RLP       6      NA
#> 3        Bigmouth Buffalo  metric    linear  75      TL    RLP     150      NA
#> 4 Bigmouth Sleepers (all)  metric quadratic  25      TL    EmP      70     390
#> 5 Bigmouth Sleepers (all)  metric quadratic  50      TL    EmP      70     390
#> 6 Bigmouth Sleepers (all)  metric quadratic  75      TL    EmP      70     390
#>      int slope   quad                  source
#> 1 -3.801 1.783  0.329 Giannetto et al. (2012)
#> 2 -3.346 3.118     NA    Bister et al. (2000)
#> 3 -5.069 3.118     NA    Bister et al. (2000)
#> 4 -6.000 3.631 -0.111  Cooney and Kwak (2010)
#> 5 -5.400 3.153 -0.011  Cooney and Kwak (2010)
#> 6 -4.323 2.237  0.188  Cooney and Kwak (2010)
#>                                                        comment
#> 1                                 not in Neumann et al. (2012)
#> 2                                                         none
#> 3                                                         none
#> 4                                 not in Neumann et al. (2012)
#> 5 Quadratic term not significant; not in Neumann et al. (2012)
#> 6                                 not in Neumann et al. (2012)