WSlit {FSA}	R Documentation

All known standard weight equations.

Description

Parameters for all known standard weight equations.

Format

A data frame with observations on the following 13 variables:

species

Species name.

units

Units of measurements. Metric uses lengths in mm and weight in grams. English uses lengths in inches and weight in pounds.

type

Type of equation (linear or quadratic).

ref

Reference quartile (75, 50, or 25).

measure

The type of length measurement used – total length (TL) or fork length (FL).

method

The type of method used to derive the equation (RLP,EmP, or Other).

min.len

Minimum total length (mm or in, depending on units) for which the equation should be applied.

max.len

Maximum total length (mm or in, depending on units) for which the equation should be applied.

int

The intercept for the model.

slope

The slope for the linear models or the linear coefficient for the quadratic equation.

quad

The quadratic coefficient in the quadratic equations.

source

Source of the equation. These match the sources given in Neumann et al. 2012.

comment

Comments about use of equation.

Details

The minimum TL for the English units were derived by rounding the converted minimum TL for the metric units to what seemed like common units (inches, half inches, or quarter inches).

Topic(s)

Relative weight

Standard weight

Condition

IFAR Chapter

8-Condition.

Source

Most of these equations can be found in Neumann, R.M., C.S. Guy, and D.W. Willis. 2012. Length, Weight, and Associated Indices. Chapter 14 in Zale, A.V., D.L. Parrish, and T.M. Sutton, editors. Fisheries Techniques. American Fisheries Society, Bethesda, MD.

Some species were not in Neumann et al (2012) and are noted as such in the comments variable.

References

Ogle, D.H. 2016. Introductory Fisheries Analyses with R. Chapman & Hall/CRC, Boca Raton, FL.

See Also

See wsVal and wrAdd for related functionality.

Examples

str(WSlit)
head(WSlit)

[Package FSA version 0.8.23 Index]