WhitefishLC {FSA}	R Documentation

Assigned ages from two readers on three structures for Lake Whitefish from Lake Champlain.

Description

Assigned ages from two readers on three structures for Lake Whitefish (Coregonus clupeaformis) from Lake Champlain in 2009.

Format

A data frame with 151 observations on the following 11 variables:

fishID

A unique fish identification number

tl

Total length (in mm)

scale1

Assessed age from scales by first reader

scale2

Assessed age from scales by second reader

scaleC

Consensus age from scales by both reader

finray1

Assessed age from fin rays by first reader

finray2

Assessed age from fin rays by second reader

finrayC

Consensus age from fin rays by both reader

otolith1

Assessed age from otoliths by first reader

otolith2

Assessed age from otoliths by second reader

otolithC

Consensus age from otoliths by both reader

Topic(s)

Age

Ageing Error

Precision

Bias

Age Comparisons

Source

Data from Herbst, S.J. and J.E. Marsden. 2011. Comparison of precision and bias of scale, fin ray, and otolith age estimates for lake whitefish (Coregonus clupeaformis) in Lake Champlain. Journal of Great Lakes Research. 37:386-389. Contributed by Seth Herbst. Do not use for other than educational purposes without permission from the author. [Was (is?) from http://www.uvm.edu/rsenr/emarsden/documents/Herbst%20and%20Marsden%20whitefish%20age%20structure%20comparison.pdf.]

See Also

Used in ageBias and agePrecision examples.

Examples

str(WhitefishLC)
head(WhitefishLC)

[Package FSA version 0.8.23 Index]