fishR {FSA}	R Documentation

Opens web pages associated with the fishR website.

Description

Opens web pages associated with the fishR website in a browser. The user can open the main page or choose a specific page to open.

Usage

fishR(where = c("home", "IFAR", "general", "books", "AIFFD", "posts",
  "news"))

Arguments

where
A string that indicates a particular page on the fishR website to open.

Value

None, but a webpage will be opened in the default browser.

Author(s)

Derek H. Ogle, derek@derekogle.com

Examples

## Not run: 
## Opens an external webpage ... only run interactively
fishR()            # home page
fishR("IFAR")      # Introduction to Fisheries Analysis with R page
fishR("general")   # examples page
fishR("books")     # examples page
fishR("AIFFD")     # Analysis & Interpretation of Freshw. Fisher. Data page
fishR("posts")     # blog posts (some examples) page

## End(Not run)

[Package FSA version 0.8.23 Index]