hoCoef {FSA}	R Documentation

Performs a hypothesis test that a linear model parameter is equal to a specific value.

Description

Performs a hypothesis test that a linear model parameter is equal to a specific value. Useful for testing that a parameter is equal to a value other than 0.

Usage

hoCoef(object, term = 2, bo = 0, alt = c("two.sided", "less",
  "greater"))

Arguments

object
A lm object.

term
A single numeric that indicates which term in the model to use in the hypothesis test.

bo
The null hypothesized parameter value.

alt
A string that identifies the “direction” of the alternative hypothesis. The strings may be "less" for a “less than” alternative, "greater" for a “greater than” alternative, or "two.sided" (DEFAULT) for a “not equals” alternative.

Details

The “direction” of the alternative hypothesis is identified by a string in the alt argument.

If the lm object is from a simple linear regression with an intercept then term=1 will use the intercept and term=2 will use the slope in the hypothesis test.

Value

A matrix that contains the term number, hypothesized value, parameter estimate, standard error of the parameter estimate, t test statistic, degrees-of-freedom, and corresponding p-value.

Author(s)

Derek H. Ogle, derek@derekogle.com

See Also

htest.nlsBoot.

Examples

# Simple linear regression test HA:slope!=0.1
lm1 <- lm(mirex~weight, data=Mirex)
hoCoef(lm1,2,0.1)

[Package FSA version 0.8.25 Index]