poiCI {FSA}	R Documentation

Confidence interval for Poisson counts.

Description

Computes a confidence interval for the Poisson counts.

Usage

poiCI(x, conf.level = 0.95, type = c("exact", "daly", "byar",
  "asymptotic"), verbose = FALSE)

Arguments

x
A single number or vector that represents the number of observed successes.

conf.level
A number that indicates the level of confidence to use for constructing confidence intervals (default is 0.95).

type
A string that identifies the type of method to use for the calculations. See details.

verbose
A logical that indicates whether x should be included in the returned matrix (=TRUE) or not (=FALSE; DEFAULT).

Details

Computes a CI for the Poisson counts using the exact, gamma distribution (daly‘), Byar’s (byar), or normal approximation (asymptotic) methods. This is largely a wrapper to pois.exact, pois.daly, pois.byar, and pois.approx functions documented in pois.conf.intin epitools.

Value

A #x2 matrix that contains the lower and upper confidence interval bounds as columns and, if verbose=TRUE x.

Author(s)

Derek H. Ogle, derek@derekogle.com

See Also

See pois.exact, pois.daly, pois.byar, and pois.approx (documented in pois.conf.int) in epitools for more description and references.

Examples

## Demonstrates using all types at once
poiCI(12)

## Selecting types
poiCI(12,type="daly")
poiCI(12,type="byar")
poiCI(12,type="asymptotic")
poiCI(12,type="asymptotic",verbose=TRUE)
poiCI(12,type=c("exact","daly"))
poiCI(12,type=c("exact","daly"),verbose=TRUE)

## Demonstrates use with multiple inputs
poiCI(c(7,10),type="exact")
poiCI(c(7,10),type="exact",verbose=TRUE)

[Package FSA version 0.8.23 Index]