Finds Gabelhouse lengths (for PSD calculations) for a species.

Returns a vector with the five Gabelhouse lengths for a chosen species.

Usage

psdVal(
  species = "List",
  group = NULL,
  units = c("mm", "cm", "in"),
  addLens = NULL,
  addNames = NULL,
  incl.zero = TRUE,
  showJustSource = FALSE
)

Arguments

species

A string that contains the species name for which to find Gabelhouse lengths. See details.

group

A string that contains the sub-group of `species` for which to find the Gabelhouse lengths. Will be things like “"landlocked"”, “"lentic"”.

units

A string that indicates the units for the returned lengths. Choices are mm for millimeters (DEFAULT), cm for centimeters, and in for inches.

addLens

A numeric vector that contains minimum length definitions for additional categories. See details.

addNames

A string vector that contains names for the additional length categories added with addLens. See details.

incl.zero

A logical that indicates if a zero is included in the first position of the returned vector (DEFAULT) or not. This position will be named “substock”. See details.

showJustSource

A logical that indicates whether just the literature source information should be returned (TRUE) or not. If TRUE this will NOT return any of the Gabelhouse length information.

Value

A vector of minimum values for length categories for the chosen species.

Details

Finds the Gabelhouse lengths from data(PSDlit) for the species given in species. The species name must be spelled exactly (within capitalization differences) as it appears in data(PSDlit). Type psdVal() to see the list of species and how they are spelled.

A zero is included in the first position of the returned vector if incl.zero=TRUE. This is useful when computing PSD values with a data.frame that contains fish smaller than the stock length.

Additional lengths may be added to the returned vector with addLens. Names for these lengths can be included in addNames. If addNames is non-NULL, then it must be of the same length as addLens. If addLens is non-NULL but addNames is NULL, then the default names will be the same as the lengths in addLens. The addLens argument is useful for calculating PSD values that are different from the Gabelhouse lengths.

IFAR Chapter

6-Size Structure.

References

Ogle, D.H. 2016. Introductory Fisheries Analyses with R. Chapman & Hall/CRC, Boca Raton, FL.

Guy, C.S., R.M. Neumann, and D.W. Willis. 2006. New terminology for proportional stock density (PSD) and relative stock density (RSD): proportional size structure (PSS). Fisheries 31:86-87. [Was (is?) from http://pubstorage.sdstate.edu/wfs/415-F.pdf.]

Guy, C.S., R.M. Neumann, D.W. Willis, and R.O. Anderson. 2006. Proportional size distribution (PSD): A further refinement of population size structure index terminology. Fisheries 32:348. [Was (is?) from http://pubstorage.sdstate.edu/wfs/450-F.pdf.]

Willis, D.W., B.R. Murphy, and C.S. Guy. 1993. Stock density indices: development, use, and limitations. Reviews in Fisheries Science 1:203-222. [Was (is?) from http://web1.cnre.vt.edu/murphybr/web/Readings/Willis%20et%20al.pdf.]

See also

Author

Derek H. Ogle, DerekOgle51@gmail.com

Examples

#===== List all available species in PSDlit
psdVal()
#> 
#> Species name must be one of following. Be careful of spelling and capitalization.
#>  [1] "Arctic Grayling"           "Bighead Carp"             
#>  [3] "Bigmouth Buffalo"          "Black Bullhead"           
#>  [5] "Black Carp"                "Black Crappie"            
#>  [7] "Blue Catfish"              "Bluegill"                 
#>  [9] "Brook Trout"               "Brown Bullhead"           
#> [11] "Brown Trout"               "Bull Trout"               
#> [13] "Burbot"                    "Chain Pickerel"           
#> [15] "Channel Catfish"           "Chinook Salmon"           
#> [17] "Common Carp"               "Cutthroat Trout"          
#> [19] "Flathead Catfish"          "Freshwater Drum"          
#> [21] "Gizzard Shad"              "Golden Trout"             
#> [23] "Goldeye"                   "Grass Carp"               
#> [25] "Green Sunfish"             "Kokanee"                  
#> [27] "Lake Chubsucker"           "Lake Trout"               
#> [29] "Largemouth Bass"           "Longnose Gar"             
#> [31] "Muskellunge"               "Northern Pike"            
#> [33] "Northern Snakehead"        "Paddlefish"               
#> [35] "Pallid Sturgeon"           "Palmetto Bass"            
#> [37] "Pumpkinseed"               "Rainbow Trout"            
#> [39] "Redbreast Sunfish"         "Redear Sunfish"           
#> [41] "River Carpsucker"          "Rock Bass"                
#> [43] "Ruffe"                     "Sauger"                   
#> [45] "Saugeye"                   "Shoal Bass"               
#> [47] "Shorthead Redhorse"        "Silver Carp"              
#> [49] "Smallmouth Bass"           "Smallmouth Buffalo "      
#> [51] "Splake"                    "Spotted Bass"             
#> [53] "Spotted Gar"               "Spotted Sunfish"          
#> [55] "Striped Bass"              "Striped Bass X White Bass"
#> [57] "Suwannee Bass"             "Utah Chub"                
#> [59] "Walleye"                   "Warmouth"                 
#> [61] "White Bass"                "White Catfish"            
#> [63] "White Crappie"             "White Perch"              
#> [65] "White Sucker"              "Yellow Bass"              
#> [67] "Yellow Bullhead"           "Yellow Perch"             

#===== Typical usages
psdVal("Bluegill")
#>  substock     stock   quality preferred memorable    trophy 
#>         0        80       150       200       250       300 
psdVal("Bluegill",units="in")
#>  substock     stock   quality preferred memorable    trophy 
#>         0         3         6         8        10        12 
psdVal("Bluegill",units="in",incl.zero=FALSE)
#>     stock   quality preferred memorable    trophy 
#>         3         6         8        10        12 
psdVal("Bluegill",showJustSource=TRUE)
#>    species             source
#> 8 Bluegill Gabelhouse (1984a)

#===== For species that have sub-groups
psdVal("Brown Trout",group="lentic")
#>  substock     stock   quality preferred memorable    trophy 
#>         0       200       300       400       500       600 
psdVal("Brown Trout",group="lotic")
#>  substock     stock   quality preferred memorable    trophy 
#>         0       150       230       300       380       460 
psdVal("Palmetto Bass",group="revised")
#>  substock     stock   quality preferred memorable    trophy 
#>         0       250       410       510       610       710 
psdVal("Palmetto Bass",group="original")
#>  substock     stock   quality preferred memorable    trophy 
#>         0       200       300       380       510       630 

#===== Adding user-defined categories
#-----   with lengths and names separately in addLens= and addNames=
psdVal("Bluegill",units="in",addLens=7)
#>  substock     stock   quality         7 preferred memorable    trophy 
#>         0         3         6         7         8        10        12 
psdVal("Bluegill",units="in",addLens=7,addNames="MinLen")
#>  substock     stock   quality    MinLen preferred memorable    trophy 
#>         0         3         6         7         8        10        12 
psdVal("Bluegill",units="in",addLens=c(7,9),addNames=c("MinSlot","MaxSlot"))
#>  substock     stock   quality   MinSlot preferred   MaxSlot memorable    trophy 
#>         0         3         6         7         8         9        10        12 
#-----   with a named vector in addLens=
psdVal("Bluegill",units="in",addLens=c("MinLen"=7))
#>  substock     stock   quality    MinLen preferred memorable    trophy 
#>         0         3         6         7         8        10        12 
psdVal("Bluegill",units="in",addLens=c("MinSlot"=7,"MaxSlot"=9))
#>  substock     stock   quality   MinSlot preferred   MaxSlot memorable    trophy 
#>         0         3         6         7         8         9        10        12