rSquared {FSA}	R Documentation

Extract the coefficient of determination from a linear model object.

Description

Extracts the coefficient of determination (i.e., “r-squared”) from a linear model (i.e., lm) object.

Usage

rSquared(object, ...)

## Default S3 method:
rSquared(object, ...)

## S3 method for class 'lm'
rSquared(object, digits = getOption("digits"),
  percent = FALSE, ...)

Arguments

object
An object saved from lm.

...
Additional arguments for methods.

digits
A single number that is the number of digits to round the returned result to.

percent
A logical that indicates if the result should be returned as a percentage (=TRUE) or as a proportion (=FALSE; default).

Details

This is a convenience function to extract the r.squared part from summary(lm).

Value

A numeric, as either a proportion or percentage, that is the coefficient of determination for a linear model.

Examples

lm1 <- lm(mirex~weight, data=Mirex)
rSquared(lm1)
rSquared(lm1,digits=3)
rSquared(lm1,digits=1,percent=TRUE)

## rSquared only works with lm objects
## Not run: 
nls1 <- nls(mirex~a*weight^b,data=Mirex,start=list(a=1,b=1))
rSquared(nls1)

## End(Not run)

[Package FSA version 0.8.23 Index]