Creates a one- or two-way table of summary statistics.

Creates a one- or two-way table of summary statistics for a quantitative variable.

sumTable(formula, ...)

# S3 method for formula
sumTable(formula, data = NULL, FUN = mean,
  digits = getOption("digits"), ...)

Arguments

formula
A formula with a quantitative variable on the left-hand-side and one or two factor variables on the right-hand-side. See details.

…
Other arguments to pass through to FUN.

data
An optional data frame that contains the variables in formula.

FUN
A scalar function that identifies the summary statistics. Applied to the quantitative variable for all data subsets identified by the combination of the factor(s). Defaults to mean.

digits
A single numeric that indicates the number of digits to be used for the result.

Value

A one-way array of values if only one factor variable is supplied on the right-hand-side of formula. A two-way matrix of values if two factor variables are supplied on the right-hand-side of formula. These are the same classes of objects returned by tapply.

Details

The formula must be of the form quantitative~factor or quantitative~factor*factor2 where quantitative is the quantitative variable to construct the summaries for and factor and factor2 are factor variables that contain the levels for which separate summaries should be constructed. If the variables on the right-hand-side are not factors, then they will be coerced to be factors and a warning will be issued.

This function is largely a wrapper to tapply(), but only works for one quantitative variable on the left-hand-side and one or two factor variables on the right-hand-side. Consider using tapply for situations with more factors on the right-hand-side.

See also

Examples

## The same examples as in the old aggregate.table in gdata package
## but data in data.frame to illustrate formula notation
d <- data.frame(g1=sample(letters[1:5], 1000, replace=TRUE),
                g2=sample(LETTERS[1:3], 1000, replace=TRUE),
                dat=rnorm(1000))

sumTable(dat~g1*g2,data=d,FUN=length)       # get sample size

#>    A  B  C
#> a 70 69 76
#> b 63 62 75
#> c 55 75 60
#> d 74 81 77
#> e 69 41 53

sumTable(dat~g1*g2,data=d,FUN=validn)       # get sample size (better way)

#>    A  B  C
#> a 70 69 76
#> b 63 62 75
#> c 55 75 60
#> d 74 81 77
#> e 69 41 53

sumTable(dat~g1*g2,data=d,FUN=mean)         # get mean

#>            A          B          C
#> a -0.0072019 -0.0532666  0.0030068
#> b -0.1359371 -0.0796800 -0.1447992
#> c  0.0143723 -0.0740245 -0.1740726
#> d -0.0747419 -0.0920961  0.1173423
#> e  0.0824972  0.2652187  0.1835084

sumTable(dat~g1*g2,data=d,FUN=sd)           # get sd

#>           A         B         C
#> a 1.0505666 0.9207278 1.0166821
#> b 1.0493635 0.9655138 0.9868420
#> c 1.0115367 1.0200775 1.0550774
#> d 0.8633723 1.0313599 1.0002891
#> e 1.0517470 1.1070658 0.9213085

sumTable(dat~g1*g2,data=d,FUN=sd,digits=1)  # show digits= argument

#>     A   B   C
#> a 1.1 0.9 1.0
#> b 1.0 1.0 1.0
#> c 1.0 1.0 1.1
#> d 0.9 1.0 1.0
#> e 1.1 1.1 0.9

## Also demonstrate use in the 1-way example -- but see Summarize() in FSA package
sumTable(dat~g1,data=d,FUN=validn)

#>   a   b   c   d   e 
#> 215 200 190 232 163 

sumTable(dat~g1,data=d,FUN=mean)

#>          a          b          c          d          e 
#> -0.0183768 -0.1218207 -0.0800301 -0.0170488  0.1613019 

## Example with a missing value (compare to above)
d$dat[1] <- NA
sumTable(dat~g1,data=d,FUN=validn)  # note use of validn

#>   a   b   c   d   e 
#> 215 199 190 232 163 

sumTable(dat~g1,data=d,FUN=mean,na.rm=TRUE)

#>          a          b          c          d          e 
#> -0.0183768 -0.1315083 -0.0800301 -0.0170488  0.1613019