Finds standard weight equation coefficients for a particular species.

Returns a vector that contains all known or a subset of information about the standard weight equation for a given species, type of measurement units, and reference percentile.

wsVal(species = "List", units = c("metric", "English"), ref = 75,
  simplify = FALSE)

Arguments

species
A string that contains the species name for which to find coefficients. See details.

units
A string that indicates whether the coefficients for the standard weight equation to be returned are in ("metric" (DEFAULT; mm and g) or "English" (in and lbs) units.

ref
A numeric that indicates which percentile the equation should be returned for. Note that the vast majority of equations only exist for the 75th percentile (DEFAULT).

simplify
A logical that indicates whether the ‘units’, ‘ref’, ‘measure’, ‘method’, ‘comments’, and ‘source’ fields should be included (=FALSE) or not (=TRUE; DEFAULT). See details.

Value

A one row data frame from WSlit that contains all known information about the standard weight equation for a given species, type of measurement units, and reference percentile if simplify=FALSE. If simplify=TRUE then only the species; minimum and maximum length for which the standard equation should be applied; and intercept, slope, and quadratic coefficients for the standard weight equation. Note that the maximum length and the quadratic coefficient will not be returned if they do not exist in WSlit.

If no arguments are given to this function, a species name is mis-spelled, or if a standard weight equation does not exist (in WSlit) for a particular species, then a warning will be issued and a list of species names will be printed.

Details

This function extract all known information from WSlit about the following standard weight equation,

$$log_{10}(Ws) = log_{10}(a) + blog_{10}(L) + blog_{10}(L)^{2}$$

See WSlit for more information about the meaning of each value returned.

Note from above that the coefficients are returned for the TRANSFORMED model. Thus, to obtain the standard weight (Ws), the returned coefficients are used to compute the common log of Ws which must then bed raised to the power of 10 to compute the Ws.

IFAR Chapter

8-Condition.

References

Ogle, D.H. 2016. Introductory Fisheries Analyses with R. Chapman & Hall/CRC, Boca Raton, FL.

See also

Examples

wsVal()

#> 
#> Species name must be one of following. Be careful of spelling and capitalization.

#>  [1] "Aegean Chub"               "Bigmouth Buffalo"         
#>  [3] "Bigmouth Sleepers (all)"   "Bigmouth Sleepers (lotic)"
#>  [5] "Black Bullhead"            "Black Crappie"            
#>  [7] "Blacktail Redhorse"        "Blue Catfish"             
#>  [9] "Blue Sucker"               "Bluegill"                 
#> [11] "Bridgelip Sucker"          "Brook Chub"               
#> [13] "Brook Trout"               "Brown Bullhead"           
#> [15] "Brown Trout (lentic)"      "Brown Trout (lotic)"      
#> [17] "Bull Trout"                "Burbot"                   
#> [19] "Cavedano chub"             "Chain Pickerel"           
#> [21] "Channel Catfish"           "Chinook Salmon"           
#> [23] "Cisco"                     "Common Carp"              
#> [25] "Cutthroat Trout (lentic)"  "Cutthroat Trout (lotic)"  
#> [27] "European chub"             "European Perch"           
#> [29] "Flannelmouth Sucker"       "Flathead Catfish"         
#> [31] "Freshwater Drum"           "Gizzard Shad"             
#> [33] "Golden Shiner"             "Golden Trout"             
#> [35] "Green Sunfish"             "Humpback Chub"            
#> [37] "Kokanee"                   "Lake herring"             
#> [39] "Lake Trout"                "Largemouth Bass"          
#> [41] "Largescale Sucker"         "Longnose Gar"             
#> [43] "Marble Trout"              "Mountain Mullet"          
#> [45] "Mountain Whitefish"        "Muskellunge (female)"     
#> [47] "Muskellunge (male)"        "Muskellunge (overall)"    
#> [49] "Northern Pike"             "Northern Pikeminnow"      
#> [51] "Northern Squawfish"        "Paddlefish (female)"      
#> [53] "Paddlefish (male)"         "Paddlefish (overall)"     
#> [55] "Palmetto Bass"             "Pejerrey"                 
#> [57] "Pumpkinseed"               "Pursak chub"              
#> [59] "Rainbow Trout (lentic)"    "Rainbow Trout (lotic)"    
#> [61] "Razorback Sucker"          "Redear Sunfish"           
#> [63] "Riffle Dace"               "River Carpsucker"         
#> [65] "River Goby"                "Rock Bass"                
#> [67] "Roundtail Chub"            "Ruffe"                    
#> [69] "Sardine"                   "Sauger"                   
#> [71] "Saugeye"                   "Shorthead Redhorse"       
#> [73] "Shovelnose Sturgeon"       "Smallmouth Bass"          
#> [75] "Smallmouth Buffalo"        "Spotted Bass"             
#> [77] "Spotted Gar"               "Striped Bass"             
#> [79] "Suwannee Bass"             "Tiger Muskellunge"        
#> [81] "Walleye"                   "Walleye (30-149 mm)"      
#> [83] "Warmouth"                  "White Bass"               
#> [85] "White Catfish"             "White Crappie"            
#> [87] "White Perch"               "White Sturgeon"           
#> [89] "White Sucker"              "Yellow Bass"              
#> [91] "Yellow Bullhead"           "Yellow Perch"             

wsVal("Bluegill")

#>     species  units   type ref measure method min.TL    int slope         source
#> 20 Bluegill metric linear  75      TL  Other     80 -5.374 3.316 Hillman (1982)

wsVal("Bluegill",units="metric")

#>     species  units   type ref measure method min.TL    int slope         source
#> 20 Bluegill metric linear  75      TL  Other     80 -5.374 3.316 Hillman (1982)

wsVal("Bluegill",units="English")

#>     species   units   type ref measure method min.TL    int slope
#> 19 Bluegill English linear  75      TL  Other      3 -3.371 3.316
#>            source
#> 19 Hillman (1982)

wsVal("Bluegill",units="English",simplify=TRUE)

#>     species min.TL    int slope
#> 19 Bluegill      3 -3.371 3.316

wsVal("Ruffe",units="metric",simplify=TRUE)

#>     species min.TL max.TL   int slope   quad
#> 131   Ruffe     55    205 -2.58 0.621 0.6073

wsVal("Ruffe",units="metric",ref=50,simplify=TRUE)

#>     species min.TL max.TL     int  slope   quad
#> 130   Ruffe     55    205 -3.3524 1.3969 0.4054