Finds standard weight equation coefficients for a particular species.

Returns a vector that contains all known or a subset of information about the standard weight equation for a given species, type of measurement units, and reference percentile.

wsVal(
  species = "List",
  units = c("metric", "English"),
  ref = 75,
  simplify = FALSE
)

Arguments

species

A string that contains the species name for which to find coefficients. See details.

units

A string that indicates whether the coefficients for the standard weight equation to be returned are in ("metric" (DEFAULT; mm and g) or "English" (in and lbs) units.

ref

A numeric that indicates which percentile the equation should be returned for. Note that the vast majority of equations only exist for the 75th percentile (DEFAULT).

simplify

A logical that indicates whether the ‘units’, ‘ref’, ‘measure’, ‘method’, ‘comments’, and ‘source’ fields should be included (=FALSE) or not (=TRUE; DEFAULT). See details.

Value

A one row data frame from WSlit that contains all known information about the standard weight equation for a given species, type of measurement units, and reference percentile if simplify=FALSE. If simplify=TRUE then only the species; minimum and maximum length for which the standard equation should be applied; and intercept, slope, and quadratic coefficients for the standard weight equation. Note that the maximum length and the quadratic coefficient will not be returned if they do not exist in WSlit. If no arguments are given to this function, a species name is mis-spelled, or if a standard weight equation does not exist (in WSlit) for a particular species, then a warning will be issued and a list of species names will be printed.

Details

This function extract all known information from WSlit about the following standard weight equation,

$$log_{10}(Ws) = log_{10}(a) + blog_{10}(L) + blog_{10}(L)^{2}$$

See WSlit for more information about the meaning of each value returned.

Note from above that the coefficients are returned for the TRANSFORMED model. Thus, to obtain the standard weight (Ws), the returned coefficients are used to compute the common log of Ws which must then bed raised to the power of 10 to compute the Ws.

IFAR Chapter

8-Condition.

References

Ogle, D.H. 2016. Introductory Fisheries Analyses with R. Chapman & Hall/CRC, Boca Raton, FL.

See also

Author

Derek H. Ogle, derek@derekogle.com

Examples

wsVal()
#> 
#> Species name must be one of following. Be careful of spelling and capitalization.
#>   [1] "Aegean Chub"                       "African Sharptooth Catfish"       
#>   [3] "Alabama Bass"                      "Ankara Nase"                      
#>   [5] "Arctic Grayling"                   "Bigmouth Buffalo"                 
#>   [7] "Bigmouth Sleepers (all)"           "Bigmouth Sleepers (lotic)"        
#>   [9] "Black Bullhead"                    "Black Crappie"                    
#>  [11] "Blacktail Redhorse"                "Blue Catfish"                     
#>  [13] "Blue Sucker"                       "Bluegill"                         
#>  [15] "Bridgelip Sucker"                  "Brook Chub"                       
#>  [17] "Brook Trout"                       "Brook Trout (appalachia)"         
#>  [19] "Brown Bullhead"                    "Brown Trout (lentic)"             
#>  [21] "Brown Trout (lotic)"               "Bull Trout"                       
#>  [23] "Burbot"                            "Cavedano Chub"                    
#>  [25] "Chain Pickerel"                    "Channel Catfish"                  
#>  [27] "Chinook Salmon"                    "Cisco"                            
#>  [29] "Common Carp"                       "Cutthroat Trout (lentic)"         
#>  [31] "Cutthroat Trout (lotic)"           "European Chub"                    
#>  [33] "European Perch"                    "Flannelmouth Sucker"              
#>  [35] "Flathead Catfish"                  "Fourbarbel Scraper"               
#>  [37] "Freshwater Drum"                   "Gizzard Shad"                     
#>  [39] "Golden Shiner"                     "Golden Trout"                     
#>  [41] "Green Sunfish"                     "Horse Barbel"                     
#>  [43] "Humpback Chub"                     "Kokanee"                          
#>  [45] "Lake Herring"                      "Lake Trout"                       
#>  [47] "Largemouth Bass"                   "Largescale Sucker"                
#>  [49] "Longnose Gar"                      "Marble Trout"                     
#>  [51] "Mountain Mullet"                   "Mountain Whitefish"               
#>  [53] "Muskellunge (female)"              "Muskellunge (male)"               
#>  [55] "Muskellunge (overall)"             "Nile Tilapia"                     
#>  [57] "Nipple-Lip Scraper"                "Northern Pike"                    
#>  [59] "Northern Pikeminnow"               "Northern Squawfish"               
#>  [61] "Paddlefish (female)"               "Paddlefish (male)"                
#>  [63] "Paddlefish (overall)"              "Palmetto Bass"                    
#>  [65] "Pejerrey"                          "Pumpkinseed"                      
#>  [67] "Pursak Chub"                       "Rainbow Trout (lentic)"           
#>  [69] "Rainbow Trout (lotic)"             "Razorback Sucker"                 
#>  [71] "Redear Sunfish"                    "Riffle Dace"                      
#>  [73] "River Carpsucker"                  "River Goby"                       
#>  [75] "Rock Bass"                         "Roundtail Chub"                   
#>  [77] "Ruffe"                             "Sardine"                          
#>  [79] "Sauger"                            "Saugeye"                          
#>  [81] "Shoal Bass"                        "Shorthead Redhorse"               
#>  [83] "Shovelnose Sturgeon"               "Smallmouth Bass"                  
#>  [85] "Smallmouth Buffalo"                "South European Roach"             
#>  [87] "Spotted Bass"                      "Spotted Bass (alabama subspecies)"
#>  [89] "Spotted Gar"                       "Striped Bass"                     
#>  [91] "Striped Bass (hybrid)"             "Striped Bass X White Bass"        
#>  [93] "Suwannee Bass"                     "Tiger Muskellunge"                
#>  [95] "Utah Chub"                         "Walleye"                          
#>  [97] "Walleye (30-149 mm)"               "Warmouth"                         
#>  [99] "White Bass"                        "White Catfish"                    
#> [101] "White Crappie"                     "White Perch"                      
#> [103] "White Sturgeon"                    "White Sucker"                     
#> [105] "Yellow Bass"                       "Yellow Bullhead"                  
#> [107] "Yellow Perch"                     
wsVal("Bluegill")
#>     species  units   type ref measure method min.TL    int slope         source
#> 26 Bluegill metric linear  75      TL  Other     80 -5.374 3.316 Hillman (1982)
wsVal("Bluegill",units="metric")
#>     species  units   type ref measure method min.TL    int slope         source
#> 26 Bluegill metric linear  75      TL  Other     80 -5.374 3.316 Hillman (1982)
wsVal("Bluegill",units="English")
#>     species   units   type ref measure method min.TL    int slope
#> 25 Bluegill English linear  75      TL  Other      3 -3.371 3.316
#>            source
#> 25 Hillman (1982)
wsVal("Bluegill",units="English",simplify=TRUE)
#>     species min.TL    int slope
#> 25 Bluegill      3 -3.371 3.316
wsVal("Ruffe",units="metric",simplify=TRUE)
#>     species min.TL max.TL   int slope   quad
#> 142   Ruffe     55    205 -2.58 0.621 0.6073
wsVal("Ruffe",units="metric",ref=50,simplify=TRUE)
#>     species min.TL max.TL     int  slope   quad
#> 141   Ruffe     55    205 -3.3524 1.3969 0.4054